Switched from vector to list for some variables
I do not need all the objects and data in contiguous blocks of memory,
and I don't need to be able to randomly access them. I much prefer
faster forward/backwards traversal and insertion/deletion so changing
from std::vector to std::list seems like a good idea.

Hopefully this is the correct choice for me.