Simplify functions and implement chapter download
Previous the function to fetch data from the mangadex chapter API was a
bit messy and overcomplicated a few things, same with the function to
download images from the chapter. So I cleaned them up a tad so all you
need to do select chapter ID and the JSON will be downloaded and parsed,
then the images will then be downloaded

Code is still very brittle and can break when the JSON response is
lacking some info or has some info as NULL, but overall it's a step in
the right direction.