Mangadex: Update to the v2 API
v2 API is overall nicer to use and work with. And with the update to it
I can now drop (in a later commit) the string maniplation stuff that I
needed to with the v1 API to split the authors on a token (In this case,
it was ',')

There is still some gotchas I need to account for that I also had
planned to account for in the v1 API, but overall there should be less
edge cases I need to try and account for at a later date. A few (but not
all) of them have been documented as comments in the header file.

Also moved to signed integers, since only unsigned integers have defined
overflow behaviour, and I have no explicit reason for them to be
unsigned other then for the simple mindset that I was only going to be
working in the postive number range. Video that alearted me to my ill
informed thoughts about unsigned/signed differences only being that one
can store negative numbers.
https://youtu.be/yG1OZ69H_-o?t=39m16s