Fix runtime error by using system libcurl
The runtime error:
SSL certificate problem: unable to get local issuer certificate

Since I was to lazy (and by extension it wasn't an issue with my code,
but a third party's code) I did the smart thing and enabled the use of
the system libcurl. As a side effect, the CMake configuration stage is
much quicker then before.

CPR was pointed to latest commit during testing on narrowing down the
bug. I decided to leave it like so for now until a new CPR release.

An unintended side effect (either by this change or something else), is
that the third party JSON tests started building, which it never use to.
So I disabled it by default for now.