Slightly simplify mangadex.h
The end goal of mangadex.h should be the ability to just parse the API
ans store it in variables, for the program to make use of at a later
date. Moving the download related functions out of mangadex.h is a step
into the direction needed for the above goal.

prettyPrint() is the only other thing that's left to remove when the time is appropriate