Allow dependencies to properly propagate automatically
Making the manga-manager dependencies public allow any targets that include it to get all working includes
and dependencies that it, itself depends on. Also the change means the
archive (.a) files are now in the _deps/ folder instead of the lib/
folder inside the build directory.

Also allows the removal of PIC flag that was needed previous due to the
bad configuration of deps