ui: Add a basic Vulkan render
This is just the ground work!
This essentially is all the necessary Vulkan boilerplate code to display
something on the screen. It can display a colored cube, yay!

There is still more work to go into  this via follow up commits, such
as:
- Abstract it away into a class
- Comment the code
- Wire up Dear Imgui for all the UI goodies

This is currently heavily based of the Vulkan-HPP samples, which can be
found at
https://github.com/KhronosGroup/Vulkan-Hpp/tree/master/samples

The reason its being committed as is so I'll have a reference for future
uses of an all in one file that renders something. I am by no means a
competent Graphics programmer or C++ programmer so there is probably a
chance for a lot of improvements in this code.