ui: Finish off initial window class implementation
As you may see in the diff, the window class is instantiated early on
due to potential race condition issues (basically we query the size of
the window before its fully created/presented), I am just noting the
potential issue here in case we accidentally lose the comment in the
source code.

It's all pretty straight forward, mostly just cut and paste from my
existing code into the class. Making it easier in the future if we want
to replace or remove SDL2.