ui: Const all the things!
Yes, most of the things that where made const can be considered a
shallow const instead of a deep const. Below is a comment copy+paste'd
from my code that explains why I did so (Which I believe and hope is
accurate).

Comment from code:
"Note: You'll see alot of const on vulkan.hpp objects, but since it's a C++
wrapper around C code the vulkan.hpp objects are const but the stuff they
wrap are still mutuble, this is known as shallow const (Deep const is where
an object and any children/objects it contains are immutable).
For some developers they may wonder why even mark it as const, in this case
its more like a contract between me (developer) and the code where I
"promise" that I wont be modifying the vulkan.hpp object *myself* after
creation. But I am fully aware that the library I am interacting with itself
may modify the content contained inside."