ui: Move Vulkan defines to build system
Move the Vulkan defines to the build system so it can be more easily
applied to source code as well as being less brittle overall.

There is also a typo fix slipped in here 👀

FindWayland.cmake was a pain to create, limited examples online, cryptic
documentation and when you do find examples online it uses some
deprecated options that makes you second guess yourself. Also lets not
forget the typos and minimal/cryptic error messages (Which you have to
search for in sea of error messages from cmakes own code.....).