ui: Move to RAII Vulkan-Hpp, move code into class
This took a little while to do, a mixture of life and me getting stuck
on a few issues that arose when moving from a sequentially ran file of
code to something that sorta resembles a C++ Class, there was also one
or two hiccups that the RAII version of Vulkan-Hpp brought to my
attention, but it was an otherwise fairly painless move from the
standard Vulkan-HPP to the RAII version of it.

I also implemented (aka stole) the debug callback function from
https://github.com/KhronosGroup/Vulkan-Hpp/blob/main/RAII_Samples/EnableValidationWithCallback/EnableValidationWithCallback.cpp
I'll have to see about making that code more my own and personal, but
for now it gets the job done and allows me to debug the code easier.